Refactored new_plan, and load_existing_land_use_plan
Added handling of unsaved changes to creating new plan.
